Make node factory TS and add with(Draft)Registraions
